Added the three projects for AOE delivery
