+++ springboot/openmarketConsumer/src/main/java/com/unicauca/openmarketConsumer/OpenmarketConsumerApplication.java
@@ -1,0 +1,13 @@
+package com.unicauca.openmarketConsumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OpenmarketConsumerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OpenmarketConsumerApplication.class, args);
+	}
+
+}
