+++ springboot/openmarketConsumer/src/test/java/com/unicauca/openmarketConsumer/OpenmarketConsumerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.unicauca.openmarketConsumer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OpenmarketConsumerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
