--- springboot/openmarketProducer/openmarketProducer/src/main/java/com/unicauca/openmarket/OpenmarketApplication.java
@@ -1,22 +1,0 @@
-package com.unicauca.openmarket;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import com.unicauca.openmarket.domain.service.OpenMarketRMQProducer;
-
-@SpringBootApplication
-public class OpenmarketApplication {
-	private final OpenMarketRMQProducer OMProducer;
-
-	@Autowired
-    public OpenmarketApplication(OpenMarketRMQProducer OMProducer) {
-        this.OMProducer = OMProducer;
-    }
-
-	public static void main(String[] args) {
-		SpringApplication.run(OpenmarketApplication.class, args);
-	}
-
-}
