+++ springboot/openmarketProducer/screenshoots/categories/afterDeletion.png
