+++ springboot/openmarketProducer/src/main/java/com/unicauca/openmarket/access/repository/ICategoryRepository.java
